moved basis and renamed propreties
