+++ autogalaxy/galaxy/to_inversion.py
@@ -11,7 +11,7 @@ from autogalaxy.analysis.adapt_images.adapt_images import AdaptImages
 from autogalaxy.profiles.light.linear import (
     LightProfileLinearObjFuncList,
 )
-from autogalaxy.profiles.light.basis import Basis
+from autogalaxy.profiles.basis import Basis
 from autogalaxy.profiles.light.abstract import LightProfile
 from autogalaxy.profiles.light.linear import LightProfileLinear
 from autogalaxy.galaxy.galaxy import Galaxy
@@ -141,7 +141,7 @@ class GalaxiesToInversion(AbstractToInversion):
                     if isinstance(light_profile, LightProfileLinear):
                         light_profile_list = [light_profile]
                     else:
-                        light_profile_list = light_profile.profile_list
+                        light_profile_list = light_profile.light_profile_list
                         light_profile_list = [
                             light_profile
                             for light_profile in light_profile_list
